+++ fotoskazka/app/Filament/Resources/Services/Schemas/ServiceForm.php
@@ -106,6 +106,15 @@ class ServiceForm
                             ->searchable()
                             ->columnSpanFull(),
                     ]),
+                Section::make('Видео')
+                    ->schema([
+                        Select::make('videos')
+                            ->multiple()
+                            ->relationship('videos', 'title')
+                            ->preload()
+                            ->searchable()
+                            ->columnSpanFull(),
+                    ]),
             ]);
     }
 }